Added tests for daos exceptiton

+++ Development/ChatServer/src/test/java/edu/northeastern/ccs/im/dao/GroupToUserDAOTest.java
@@ -8,6 +8,9 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.lang.reflect.Field;
+
+import edu.northeastern.ccs.im.exceptions.DatabaseConnectionException;
 import edu.northeastern.ccs.im.model.Groups;
 
 public class GroupToUserDAOTest {
@@ -15,10 +18,16 @@ public class GroupToUserDAOTest {
 	GroupToUserDAO groupToUserDAO;
 	GroupDAO groupDAO;
 	Groups group;
-	
+	boolean isException;
+
 	@Before
-	public void setUp() {
+	public void setUp() throws NoSuchFieldException, IllegalAccessException {
 		groupToUserDAO = GroupToUserDAO.getInstance();
+		Class clazz = GroupToUserDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupToUserDAO, new ConnectionManager());
+		isException = false;
 		groupDAO = GroupDAO.getInstance();
 		group = new Groups("testGroup", 2);
 		group = groupDAO.createGroup(group);
@@ -26,16 +35,39 @@ public class GroupToUserDAOTest {
 	
 	@After
 	public void destroy() {
-		groupDAO.deleteGroupByID(group.getGrpID());
+		if(!isException) {
+			groupDAO.deleteGroupByID(group.getGrpID());
+		}
+		isException = false;
 	}
 	
 	@Test
 	public void testAddUser(){
 		groupToUserDAO.addUserToGroup(2, group.getGrpID());
 	}
-	
-	@Test
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testAddUserException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupToUserDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupToUserDAO, new ConnectionTest());
+		isException = true;
+		groupToUserDAO.addUserToGroup(2, group.getGrpID());
+	}
+
+	@Test(expected = DatabaseConnectionException.class)
 	public void testAddUserFail() {
+		groupToUserDAO.addUserToGroup(0, 1);
+	}
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testAddUserFailException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupToUserDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupToUserDAO, new ConnectionTest());
+		isException = true;
 		groupToUserDAO.addUserToGroup(0, 1);
 	}
 	
@@ -43,6 +75,17 @@ public class GroupToUserDAOTest {
 	public void testCheckIfUserInGroup() {
 		groupToUserDAO.addUserToGroup(2, group.getGrpID());
 		assertTrue(groupToUserDAO.checkIfUserInGroup(2,  group.getGrpID()));
+	}
+
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testCheckIfUserInGroupException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupToUserDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupToUserDAO, new ConnectionTest());
+		isException = true;
+		groupToUserDAO.addUserToGroup(2, group.getGrpID());
 	}
 	
 	@Test
@@ -56,6 +99,18 @@ public class GroupToUserDAOTest {
 		groupToUserDAO.deleteUserFromAllGroups(1);
 		assertFalse(groupToUserDAO.checkIfUserInGroup(1, group.getGrpID()));
 	}
+
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testDeleteUserException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupToUserDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupToUserDAO, new ConnectionTest());
+		isException = true;
+		groupToUserDAO.addUserToGroup(2, group.getGrpID());
+		groupToUserDAO.deleteUserFromAllGroups(1);
+	}
 	
 	@Test
 	public void testDeleteUserFail() {
@@ -68,8 +123,20 @@ public class GroupToUserDAOTest {
 		groupToUserDAO.deleteUserFromGroup(0, group.getGrpID());
 		//assertFalse(groupToUserDAO.checkIfUserInGroup(1, group.getGrpID()));
 	}
+
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testDeleteUserFromGroupException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupToUserDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupToUserDAO, new ConnectionTest());
+		isException = true;
+		groupToUserDAO.addUserToGroup(2, group.getGrpID());
+		groupToUserDAO.deleteUserFromGroup(0, group.getGrpID());
+	}
 	
-	@Test
+	@Test(expected = DatabaseConnectionException.class)
 	public void testCreateException() {
 		groupToUserDAO.addUserToGroup(4, 5);
 	}
@@ -77,6 +144,17 @@ public class GroupToUserDAOTest {
 	@Test
 	public void testGetAllUsers() {
 		assertNotNull(groupToUserDAO.getAllUsersInGroup(group.getGrpName()));
+	}
+
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testGetAllUsersException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupToUserDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupToUserDAO, new ConnectionTest());
+		isException = true;
+		groupToUserDAO.getAllUsersInGroup(group.getGrpName());
 	}
 
 }
