+++ Development/ChatServer/src/test/java/edu/northeastern/ccs/im/dao/GroupsDAOTest.java
@@ -7,26 +7,51 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.lang.reflect.Field;
+
+import edu.northeastern.ccs.im.exceptions.DatabaseConnectionException;
 import edu.northeastern.ccs.im.model.Groups;
 
 public class GroupsDAOTest {
 	
 	GroupDAO groupDAO;
 	Groups group;
+	boolean isException;
 	
 	@Before
-	public void setUp() {
+	public void setUp() throws NoSuchFieldException, IllegalAccessException {
 		groupDAO = GroupDAO.getInstance();
+		Class clazz = GroupDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupDAO, new ConnectionManager());
+
 		group = new Groups("Group 12", 2);
+		isException = false;
 	}
 	
 	@After
 	public void destroy() {
-		groupDAO.deleteGroupByID(group.getGrpID());
+		if(!isException && groupDAO.checkGroupExists(group.getGrpName())) {
+			groupDAO.deleteGroupByID(group.getGrpID());
+		}
+		isException = false;
+
 	}
 
 	@Test
 	public void testCreateGroup() {
+		group = new Groups("GroupTest", 2);
+		groupDAO.createGroup(group);
+	}
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testCreateGroupException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupDAO, new ConnectionTest());
+		isException = true;
 		group = new Groups("GroupTest", 2);
 		groupDAO.createGroup(group);
 	}
@@ -35,11 +60,31 @@ public class GroupsDAOTest {
 	public void testDeleteGroup() {
 		groupDAO.deleteGroupByID(group.getGrpID());
 	}
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testDeleteGroupException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupDAO, new ConnectionTest());
+		isException = true;
+		groupDAO.deleteGroupByID(group.getGrpID());
+	}
 	
 	@Test
 	public void testGroupExistsID() {
 		groupDAO.checkGroupExists(group.getGrpID());
 		assertEquals("Group 12", group.getGrpName());
+	}
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testGroupExistsIDException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupDAO, new ConnectionTest());
+		isException = true;
+		groupDAO.checkGroupExists(group.getGrpID());
 	}
 	
 	@Test
@@ -48,8 +93,29 @@ public class GroupsDAOTest {
 		assertEquals("Group 12", group.getGrpName());
 	}
 
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testGroupExistsNameException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupDAO, new ConnectionTest());
+		isException = true;
+		groupDAO.checkGroupExists("Group 12");
+	}
+
 	@Test
 	public void testValidateAdmin() {
+		groupDAO.validateGroupAdmin("Group 12", UserDAO.getInstance().getUserByUserID(2).getUsername());
+	}
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testValidateAdminException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupDAO, new ConnectionTest());
+		isException = true;
 		groupDAO.validateGroupAdmin("Group 12", UserDAO.getInstance().getUserByUserID(2).getUsername());
 	}
 	
@@ -59,6 +125,17 @@ public class GroupsDAOTest {
 		Groups group1 = groupDAO.getGroupByGroupName("Group 12");
 		assertEquals(2, group1.getAdminID());
 	}
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testGroupByNameException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupDAO, new ConnectionTest());
+		isException = true;
+		groupDAO.createGroup(group);
+		Groups group1 = groupDAO.getGroupByGroupName("Group 12");
+	}
 	
 	@Test
 	public void testGroupByID() {
@@ -66,13 +143,24 @@ public class GroupsDAOTest {
 		group = groupDAO.getGroupByGroupName(group.getGrpName());
 		assertEquals(group.getGrpID(), groupDAO.getGroupByGroupID(group.getGrpID()).getGrpID());
 	}
-	
-	@Test
+
+	@Test(expected = DatabaseConnectionException.class)
+	public void testGroupByIDException() throws NoSuchFieldException, IllegalAccessException {
+		Class clazz = GroupDAO.class;
+		Field connectionManager = clazz.getDeclaredField("connectionManager");
+		connectionManager.setAccessible(true);
+		connectionManager.set(groupDAO, new ConnectionTest());
+		isException = true;
+		groupDAO.createGroup(group);
+		group = groupDAO.getGroupByGroupName(group.getGrpName());
+	}
+
+	@Test(expected = DatabaseConnectionException.class)
 	public void testGroupByIDFalse() {
 		assertNull(groupDAO.getGroupByGroupID(1));
 	}
-	
-	@Test
+
+	@Test(expected = DatabaseConnectionException.class)
 	public void testGroupByNameFalse() {
 		assertNull(groupDAO.getGroupByGroupName("x"));
 	}
